agent/pyproject.toml
[project]
name = "keiba-oracle-agent"
version = "0.1.0"
description = "Keiba Oracle - Japanese Horse Racing AI Agent with Explicit Reasoning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangGraph & LangChain Core
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # Google Generative AI SDK (Gemini 3)
    "google-genai>=1.0.0",

    # Tavily Search
    "langchain-tavily>=0.1.0",
    "tavily-python>=0.5.0",

    # FastAPI & Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # AG-UI Protocol for CopilotKit
    "copilotkit>=0.1.0",

    # Pydantic for State Models
    "pydantic>=2.10.0",

    # Environment & Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
extraPaths = [".venv/lib/python3.12/site-packages"]